manage clientid on server; refactor for queryparm
