+++ server/src/staticsites/staticsites.controller.ts
@@ -2,14 +2,20 @@ import { Controller, Post, Body, HttpException, Response, Get, Session, HttpCode
 import { ConfigService } from '../shared/config/config.service';
 import { HttpService } from '../shared/http/http.service';
 import { Constants } from '../constants';
+import { HttpUtil } from 'src/utilities/http.util';
 
-@Controller('api')
+@Controller('api/staticsites')
 export class StaticSitesController {
   constructor(private configService: ConfigService, private httpService: HttpService) {}
 
+  @Get('github/clientId')
+  clientId() {
+    return { client_id: this.configService.get('STATICSITES_GITHUB_CLIENT_ID') };
+  }
+
   @Post('github/generateAccessToken')
   @HttpCode(200)
-  async generateAccessToken(@Body('code') code: string, @Body('state') state: string, @Res() res) {
+  async generateAccessToken(@Body('code') code: string, @Body('state') state: string) {
     try {
       const response = await this.httpService.post(`${Constants.oauthApis.githubApiUri}/access_token`, {
         code,
@@ -18,7 +24,8 @@ export class StaticSitesController {
         client_secret: this.configService.get('STATICSITES_GITHUB_CLIENT_SECRET'),
       });
 
-      res.json(response.data);
+      const accessToken = HttpUtil.getQueryParameterValue('access_token', `?${response.data}`);
+      return { access_token: accessToken };
     } catch (err) {
       if (err.response) {
         throw new HttpException(err.response.data, err.response.status);
